+++ android/src/main/java/com/otahotupdate/OtaHotUpdate.kt
@@ -50,6 +50,7 @@ class OtaHotUpdate : BaseReactPackage() {
       }
     }
     fun bundleJS(context: Context): String {
+      Thread.setDefaultUncaughtExceptionHandler(CrashHandler(context))
       val sharedPrefs = SharedPrefs(context)
       val pathBundle = sharedPrefs.getString(PATH)
       val version = sharedPrefs.getString(VERSION)
