+++ android/src/main/java/com/otahotupdate/OtaHotUpdateModule.kt
@@ -1,8 +1,6 @@
 package com.otahotupdate
 
 import android.content.Context
-import android.icu.text.SimpleDateFormat
-import android.util.Log
 import com.facebook.react.bridge.Promise
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.bridge.ReactMethod
@@ -15,95 +13,12 @@ import com.rnhotupdate.Common.VERSION
 import com.rnhotupdate.Common.METADATA
 import com.rnhotupdate.SharedPrefs
 import java.io.File
-import java.util.Date
-import java.util.Locale
-import java.util.zip.ZipFile
 
 class OtaHotUpdateModule internal constructor(context: ReactApplicationContext) :
   OtaHotUpdateSpec(context) {
-
+  private val utils: Utils = Utils(context)
   override fun getName(): String {
     return NAME
-  }
-
-  private fun deleteDirectory(directory: File): Boolean {
-    if (directory.isDirectory) {
-      // List all files and directories in the current directory
-      val files = directory.listFiles()
-      if (files != null) {
-        // Recursively delete all files and directories
-        for (file in files) {
-          if (!deleteDirectory(file)) {
-            return false
-          }
-        }
-      }
-    }
-    // Finally, delete the empty directory or file
-    return directory.delete()
-  }
-  private fun deleteOldBundleIfneeded(pathKey: String?): Boolean {
-    val pathName = if (pathKey != null) pathKey else PREVIOUS_PATH
-    val sharedPrefs = SharedPrefs(reactApplicationContext)
-    val path = sharedPrefs.getString(pathName)
-    val file = File(path)
-    if (file.exists() && file.isFile) {
-      val isDeleted = deleteDirectory(file.parentFile)
-      sharedPrefs.putString(pathName, "")
-      return isDeleted
-    } else {
-      return false
-    }
-  }
-  private fun extractZipFile(
-    zipFile: File,extension: String
-  ): String? {
-    return try {
-      val outputDir = zipFile.parentFile
-      val timestamp = SimpleDateFormat("yyyyMMdd_HHmmss", Locale.getDefault()).format(Date())
-      var topLevelFolder: String? = null
-      var bundlePath: String? = null
-      ZipFile(zipFile).use { zip ->
-        zip.entries().asSequence().forEach { entry ->
-          zip.getInputStream(entry).use { input ->
-            if (topLevelFolder == null) {
-            // Get root folder of zip file after unzip
-              val parts = entry.name.split("/")
-              if (parts.size > 1) {
-                topLevelFolder = parts.first()
-              }
-            }
-            val outputFile = File(outputDir, entry.name)
-            if (entry.isDirectory) {
-              if (!outputFile.exists()) outputFile.mkdirs()
-            } else {
-              if (outputFile.parentFile?.exists() != true)  outputFile.parentFile?.mkdirs()
-              outputFile.outputStream().use { output ->
-                input.copyTo(output)
-              }
-              if (outputFile.absolutePath.endsWith(extension)) {
-                bundlePath = outputFile.absolutePath
-                return@use // Exit early if found
-              }
-            }
-          }
-        }
-      }
-      // Rename the detected top-level folder
-      if (topLevelFolder != null) {
-        val extractedFolder = File(outputDir, topLevelFolder)
-        val renamedFolder = File(outputDir, "output_$timestamp")
-        if (extractedFolder.exists()) {
-          extractedFolder.renameTo(renamedFolder)
-          // Update bundlePath if the file was inside the renamed folder
-          bundlePath = bundlePath?.replace(extractedFolder.absolutePath, renamedFolder.absolutePath)
-        }
-      }
-      bundlePath
-    } catch (e: Exception) {
-      e.printStackTrace()
-      null
-    }
   }
 
 
@@ -112,8 +27,8 @@ class OtaHotUpdateModule internal constructor(context: ReactApplicationContext) 
     if (path != null) {
       val file = File(path)
       if (file.exists() && file.isFile) {
-        deleteOldBundleIfneeded(null)
-        val fileUnzip = extractZipFile(file, extension ?: ".bundle")
+        utils.deleteOldBundleIfneeded(null)
+        val fileUnzip = utils.extractZipFile(file, extension ?: ".bundle")
         if (fileUnzip != null) {
           file.delete()
           val sharedPrefs = SharedPrefs(reactApplicationContext)
@@ -122,11 +37,14 @@ class OtaHotUpdateModule internal constructor(context: ReactApplicationContext) 
             sharedPrefs.putString(PREVIOUS_PATH, oldPath)
           }
           sharedPrefs.putString(PATH, fileUnzip)
-          sharedPrefs.putString(CURRENT_VERSION_NAME, reactApplicationContext?.getPackageInfo()?.versionName)
+          sharedPrefs.putString(
+            CURRENT_VERSION_NAME,
+            reactApplicationContext?.getPackageInfo()?.versionName
+          )
           promise.resolve(true)
         } else {
           file.delete()
-          deleteDirectory(file.parentFile)
+          utils.deleteDirectory(file.parentFile)
           val sharedPrefs = SharedPrefs(reactApplicationContext)
           sharedPrefs.putString(PATH, "")
           promise.resolve(false)
@@ -141,8 +59,8 @@ class OtaHotUpdateModule internal constructor(context: ReactApplicationContext) 
 
   @ReactMethod
   override fun deleteBundle(i: Double, promise: Promise) {
-    val isDeleted = deleteOldBundleIfneeded(PATH)
-    val isDeletedOldPath = deleteOldBundleIfneeded(PREVIOUS_PATH)
+    val isDeleted = utils.deleteOldBundleIfneeded(PATH)
+    val isDeletedOldPath = utils.deleteOldBundleIfneeded(PREVIOUS_PATH)
     val sharedPrefs = SharedPrefs(reactApplicationContext)
     sharedPrefs.putString(VERSION, "0")
     promise.resolve(isDeleted && isDeletedOldPath)
@@ -151,7 +69,7 @@ class OtaHotUpdateModule internal constructor(context: ReactApplicationContext) 
   @ReactMethod
   override fun restart() {
     val context: Context? = currentActivity
-    ProcessPhoenix.triggerRebirth(context);
+    ProcessPhoenix.triggerRebirth(context)
   }
 
   @ReactMethod
@@ -195,7 +113,10 @@ class OtaHotUpdateModule internal constructor(context: ReactApplicationContext) 
   override fun setExactBundlePath(path: String?, promise: Promise) {
     val sharedPrefs = SharedPrefs(reactApplicationContext)
     sharedPrefs.putString(PATH, path)
-    sharedPrefs.putString(CURRENT_VERSION_NAME, reactApplicationContext?.getPackageInfo()?.versionName)
+    sharedPrefs.putString(
+      CURRENT_VERSION_NAME,
+      reactApplicationContext?.getPackageInfo()?.versionName
+    )
     promise.resolve(true)
   }
 
@@ -204,7 +125,7 @@ class OtaHotUpdateModule internal constructor(context: ReactApplicationContext) 
     val sharedPrefs = SharedPrefs(reactApplicationContext)
     val oldPath = sharedPrefs.getString(PREVIOUS_PATH)
     if (oldPath != "") {
-      val isDeleted = deleteOldBundleIfneeded(PATH)
+      val isDeleted = utils.deleteOldBundleIfneeded(PATH)
       if (isDeleted) {
         sharedPrefs.putString(PATH, oldPath)
         sharedPrefs.putString(PREVIOUS_PATH, "")
