+++ example/android/app/src/main/java/otahotupdate/example/MainApplication.kt
@@ -11,6 +11,7 @@ import com.facebook.react.defaults.DefaultReactHost.getDefaultReactHost
 import com.facebook.react.defaults.DefaultReactNativeHost
 import com.facebook.react.soloader.OpenSourceMergedSoMapping
 import com.facebook.soloader.SoLoader
+import com.otahotupdate.CrashHandler
 import com.otahotupdate.OtaHotUpdate
 
 class MainApplication : Application(), ReactApplication {
